requirement.txt
python-dotenv
vk_api
requests
openai
flask